requirements.txt
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
coloredlogs==9.0
Django==1.11.5
django-oauth-toolkit==1.0.0
django-widget-tweaks==1.4.1
djangorestframework==3.7.7
humanfriendly==4.8
idna==2.6
oauthlib==2.0.6
pyreadline==2.1
pytz==2018.3
requests==2.18.4
urllib3==1.22
